+++ include/scrimmage/metrics/Metrics.h
@@ -32,7 +32,7 @@
 
 #ifndef INCLUDE_SCRIMMAGE_METRICS_METRICS_H_
 #define INCLUDE_SCRIMMAGE_METRICS_METRICS_H_
-#include <scrimmage/plugin_manager/Plugin.h>
+#include <scrimmage/entity/EntityPlugin.h>
 
 #include <map>
 #include <list>