+++ include/scrimmage/motion/MotionModel.h
@@ -36,7 +36,7 @@
 #include <Eigen/Dense>
 
 #include <scrimmage/fwd_decl.h>
-#include <scrimmage/plugin_manager/Plugin.h>
+#include <scrimmage/entity/EntityPlugin.h>
 
 #include <map>
 #include <vector>